Skip fluidfft related tests.
Using extensive use of two new decorators:
- skip_if_no_fluidfft
- classproperty

Unit tests imports were made more modular. Tests which fail when fluidfft is
not installed are skipped.

+++ fluidsim/util/console/test_profile.py
@@ -1,15 +1,13 @@
-import unittest
+import os
 import sys
+import unittest
 from glob import glob
 from shutil import rmtree
-import os
 
 from fluiddyn.util import mpi
-
-from fluidsim.util.testing import TestCase
+from fluidsim.util.testing import TestCase, skip_if_no_fluidfft
 
 from .__main__ import run_profile
-
 
 if mpi.rank == 0:
     pid = os.getpid()
@@ -25,9 +23,10 @@ path_dir = f"/tmp/tmp_test_fluidsim_profile_dir_pid{pid}"
 class TestsProfile(TestCase):
     @classmethod
     def tearDownClass(cls):
-        if mpi.rank == 0:
+        if mpi.rank == 0 and os.path.isdir(path_dir):
             rmtree(path_dir)
 
+    @skip_if_no_fluidfft()
     def test3d(self):
         command = "fluidsim-profile 8 -s ns3d -o"
         args = command.split()
@@ -44,12 +43,28 @@ class TestsProfile(TestCase):
             sys.argv = args
             run_profile()
 
+    @skip_if_no_fluidfft()
     def test2d(self):
-        command = "fluidsim-profile 16 -s ns3d -o"
+        command = "fluidsim-profile 16 -s ns2d -o"
         args = command.split()
         args.append(path_dir)
         sys.argv = args
         run_profile()
+
+    @unittest.skipIf(
+        mpi.nb_proc > 1, "Profiling solver ad1d is not meant to be done with MPI"
+    )
+    def test1d(self):
+        command = "fluidsim-profile 128 -d1 -s ad1d -o"
+        args = command.split()
+        args.append(path_dir)
+        sys.argv = args
+
+        # FIXME: works with pytest, but not with unittest
+        #  from fluidsim.util.console.util import ConsoleError
+        #  with self.assertRaises(ConsoleError):
+        #      # No profiling implemented for 1D solvers
+        #      run_profile()
 
 
 if __name__ == "__main__":
